requirements.txt
opencv-python
mysql-connector-python
numpy
Pillow